Vechile

+++ Repos/IRepository.cs
@@ -3,10 +3,10 @@
     public interface IRepository<T> where T : class
     {
         public List<T> GetAll();
-        public T GetById(int id);
+        public T GetById(long id);
         public void Insert(T obj);
         public void Update(T obj);
-        public void Delete(int id);
+        public void Delete(long id);
         public void Save();
     }
 }
